fixed strategy scheduler data-structure

+++ pheet/misc/bitops_x86.h
@@ -12,11 +12,11 @@
 namespace pheet {
 
 /* find last bit set */
-static inline unsigned int find_last_bit_set(unsigned long int x) {
+static inline unsigned int find_last_bit_set(unsigned int x) {
 if(x == 0) {
 	return 0;
 }
-unsigned long int r;
+unsigned int r;
 
 __asm__ ("bsrl %0, %1\n"
          "incl %1\n"
